rearrange software
